+++ src/components/SelectShip.tsx
@@ -1,52 +1,57 @@
-import * as React from 'react';
-import Box from '@mui/material/Box';
-import InputLabel from '@mui/material/InputLabel';
-import MenuItem from '@mui/material/MenuItem';
-import FormControl from '@mui/material/FormControl';
-import Select, { SelectChangeEvent } from '@mui/material/Select';
-import shipdata from 'data/ships/shipdata.json'
-import { CustomizedShip } from '../classes/CustomizedShip';
-import { getShip, Ship } from '../@types/Ship';
-import { NavbarStateType } from '../@types/NavbarState';
-import ListSubheader from '@mui/material/ListSubheader';
-import { NavContext } from '../App';
+import * as React from "react";
+import Box from "@mui/material/Box";
+import InputLabel from "@mui/material/InputLabel";
+import MenuItem from "@mui/material/MenuItem";
+import FormControl from "@mui/material/FormControl";
+import Select, { SelectChangeEvent } from "@mui/material/Select";
+import shipdata from "data/ships/shipdata.json";
+import { CustomizedShip } from "../classes/CustomizedShip";
+import { getShip, Ship } from "../@types/Ship";
+import { NavbarStateType } from "../@types/NavbarState";
+import ListSubheader from "@mui/material/ListSubheader";
+import { NavContext } from "../App";
 
 function getContent(header: string) {
-  var data = Object.values(shipdata)
-  switch(header) {
+  var data = Object.values(shipdata);
+  switch (header) {
     // case 'Civilian':
     //   data = data.filter(sh => sh.properties.includes('civilian'))
     //   break;
-    case 'Small':
-      data = data.filter(sh => sh.size.category == 'small')
+    case "Small":
+      data = data.filter((sh) => sh.size.category == "small");
       break;
-    case 'Medium':
-      data = data.filter(sh => sh.size.category == 'medium')
+    case "Medium":
+      data = data.filter((sh) => sh.size.category == "medium");
       break;
-    case 'Large':
-      data = data.filter(sh => sh.size.category == 'large')
+    case "Large":
+      data = data.filter((sh) => sh.size.category == "large");
       break;
-    case 'Huge':
-      data = data.filter(sh => sh.size.category == 'huge')
+    case "Huge":
+      data = data.filter((sh) => sh.size.category == "huge");
       break;
-    case 'Gargantuan':
-      data = data.filter(sh => sh.size.category == 'gargantuan')
+    case "Gargantuan":
+      data = data.filter((sh) => sh.size.category == "gargantuan");
       break;
   }
-  return data
+  return data;
 }
 
-export default function SelectShip({ ship, shipsetter, objectsetter }: { ship: string, shipsetter: Function, objectsetter: Function }) {
-  const {
-    state,
-    setter
-  } = React.useContext(NavContext) as NavbarStateType;
+export default function SelectShip({
+  ship,
+  shipsetter,
+  objectsetter,
+}: {
+  ship: string;
+  shipsetter: Function;
+  objectsetter: Function;
+}) {
+  const { state, setter } = React.useContext(NavContext) as NavbarStateType;
 
   const handleChange = (event: SelectChangeEvent) => {
     shipsetter(event.target.value as string);
-    const newship = new CustomizedShip(getShip(event.target.value) as Ship)
-    objectsetter(newship)
-    setter('')
+    const newship = new CustomizedShip(getShip(event.target.value) as Ship);
+    objectsetter(newship);
+    setter("");
   };
 
   return (
@@ -61,7 +66,7 @@ export default function SelectShip({ ship, shipsetter, objectsetter }: { ship: s
           onChange={handleChange}
           MenuProps={{
             MenuListProps: {
-              subheader: <li />
+              subheader: <li />,
             },
             PaperProps: {
               style: {
@@ -76,24 +81,34 @@ export default function SelectShip({ ship, shipsetter, objectsetter }: { ship: s
             <MenuItem key={ship.name} value={ship.name}>{ship.name}</MenuItem>
           ))} */}
           <ListSubheader>Small</ListSubheader>
-          {getContent('Small').map((ship) => (
-            <MenuItem key={ship.name} value={ship.name}>{ship.name}</MenuItem>
+          {getContent("Small").map((ship) => (
+            <MenuItem key={ship.name} value={ship.name}>
+              {ship.name}
+            </MenuItem>
           ))}
           <ListSubheader>Medium</ListSubheader>
-          {getContent('Medium').map((ship) => (
-            <MenuItem key={ship.name} value={ship.name}>{ship.name}</MenuItem>
+          {getContent("Medium").map((ship) => (
+            <MenuItem key={ship.name} value={ship.name}>
+              {ship.name}
+            </MenuItem>
           ))}
           <ListSubheader>Large</ListSubheader>
-          {getContent('Large').map((ship) => (
-            <MenuItem key={ship.name} value={ship.name}>{ship.name}</MenuItem>
+          {getContent("Large").map((ship) => (
+            <MenuItem key={ship.name} value={ship.name}>
+              {ship.name}
+            </MenuItem>
           ))}
           <ListSubheader>Huge</ListSubheader>
-          {getContent('Huge').map((ship) => (
-            <MenuItem key={ship.name} value={ship.name}>{ship.name}</MenuItem>
+          {getContent("Huge").map((ship) => (
+            <MenuItem key={ship.name} value={ship.name}>
+              {ship.name}
+            </MenuItem>
           ))}
           <ListSubheader>Gargantuan</ListSubheader>
-          {getContent('Gargantuan').map((ship) => (
-            <MenuItem key={ship.name} value={ship.name}>{ship.name}</MenuItem>
+          {getContent("Gargantuan").map((ship) => (
+            <MenuItem key={ship.name} value={ship.name}>
+              {ship.name}
+            </MenuItem>
           ))}
         </Select>
       </FormControl>