+++ Changelog.md
@@ -2,7 +2,7 @@
 - Round tonnage/engine operation time/fuel capacity to 2 decimals
 - Statblock displays magic core engine trait, automated pilot trait
 - Removed Harpoon-spear, slightly nerfed frigate AC, increased the cost of the Fluyt, added bow property to Howitzer/Mortar and prices to Titan Blades (12/1/2025 homebrew update)
-- Added Gargantuan size, and made the Man O' War Gargantuan
+- Added Gargantuan size, and made the Man O' War Gargantuan (12/1/2025 homebrew update)
 - Daily speed is now affected by armor type
 - Melee weapons have range, Titan blade has dynamic range